always pass standard in as the current list is incomplete
also, when running post-checkout, it would pass via stdin without
the use of tempfile.  did not see why.. however, with a symlink
under cygwin it was not passing any stdin as expected when reading
the code.  this issue was resolved for cygwin with the following code

                if [ "$TERM" = "cygwin" ]; then
                        [ -h "$hook" ] && hook="$(readlink -f $hook)"
                fi

but the following code was not necessary. by always performing the
tempfile stdin is passed regardless under cygwin.

Doing the tempfile stuff for every hook might be slightly slower,
but absent an exhaustive list of hooks which use STDIN, this seemed
safer.

I noticed the following hooks pass something via standard in:
prepare-commit-msg, commit-msg, post-checkout

My testing didn't display what pre-push was passing via stdin
I read that post-rewrite also passes something via stdin but
it isn't part of any of my current git workflows.